Fixing a recent change.
I wanted to pass the data by reference, but apparently I was telling it
to build an array of references.  I think since this is an array of
std::vectors, it will pass a pointer and thus effectively be pass by
reference and not do a copy.  But I'm not 100% certain.